+++ src/radios/Thetis_GPS.cpp
@@ -11,23 +11,29 @@ MicroNMEA nmea(nmeaBuffer, sizeof(nmeaBuffer));
 bool ledState = LOW;
 volatile bool ppsTriggered = false;
 
-bool initGPS(HardwareSerial &gps, Stream &out) {
-    out.print("Initializing GPS..."); // DEBUG
-    gps.begin(9600); // Begin talking with GPS at default 9600 baud.
+bool initGPS() {
+    #ifdef GPS_DEBUG
+    DEBUG_SERIAL_PORT.print("Initializing GPS..."); // DEBUG
+    #endif
+    GPS.begin(9600); // Begin talking with GPS at default 9600 baud.
     // TODO: Automatically determine GPS initial baudrate
-    if (!gps) {
-        out.println("Failed to initialize GPS"); // DEBUG
+    if (!GPS) {
+        #ifdef GPS_DEBUG
+        DEBUG_SERIAL_PORT.println("Failed to initialize GPS"); // DEBUG
+        #endif
         return false;
     }
     delay(10);
-    MicroNMEA::sendSentence(gps, "$PMTK251,38400");         // Set GPS baudrate to 38400
+    MicroNMEA::sendSentence(GPS, "$PMTK251,38400");         // Set GPS baudrate to 38400
     delay(10);
-    gps.begin(38400);
-    MicroNMEA::sendSentence(gps, "$PMTK314,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0"); // Enable NMEA RMC and GGA sentences
-    MicroNMEA::sendSentence(gps, "$PMTK220,100");           // Set GPS update rate to 100 ms (10 Hz)
+    GPS.begin(38400);
+    MicroNMEA::sendSentence(GPS, "$PMTK314,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0"); // Enable NMEA RMC and GGA sentences
+    MicroNMEA::sendSentence(GPS, "$PMTK220,100");           // Set GPS update rate to 100 ms (10 Hz)
 
     // TODO: implement a check for good GPS data
-    out.println("done!"); // DEBUG
+    #ifdef GPS_DEBUG
+    DEBUG_SERIAL_PORT.println("done!"); // DEBUG
+    #endif
     return true;
 }
 
@@ -49,14 +55,14 @@ bool initGPS(HardwareSerial &gps, Stream &out) {
 // =========================
 
 
-void getISO8601Time(char *buf, long curMillis) {
+void getISO8601Time(char *buf) {
     static long _oldMillis = 0;
     static uint8_t _oldSecond = 0;
     if (nmea.getSecond() != _oldSecond) { // One second has passed, reset millis offset and second
         _oldSecond = nmea.getSecond();
-        _oldMillis = curMillis;
+        _oldMillis = millis();
     }
-    sprintf(buf, "%04d-%02d-%02dT%02d:%02d:%02d.%03dZ",nmea.getYear(), nmea.getMonth(), nmea.getDay(), nmea.getHour(), nmea.getMinute(), nmea.getSecond(), curMillis-_oldMillis);
+    sprintf(buf, "%04d-%02d-%02dT%02d:%02d:%02d.%03dZ",nmea.getYear(), nmea.getMonth(), nmea.getDay(), nmea.getHour(), nmea.getMinute(), nmea.getSecond(), millis()-_oldMillis);
 }
 
 
@@ -69,8 +75,8 @@ void ppsHandler(void) {
     ppsTriggered = true;
 }
 
-void printUnknownSentence(MicroNMEA &nmea, Stream &out) {
-    out.println();
-	out.print("Unknown sentence: ");
-	out.println(nmea.getSentence());
+void printUnknownSentence(MicroNMEA &nmea) {
+    DEBUG_SERIAL_PORT.println();
+	DEBUG_SERIAL_PORT.print("Unknown sentence: ");
+	DEBUG_SERIAL_PORT.println(nmea.getSentence());
 }