+++ src/sensors/Thetis_BNO055.cpp
@@ -1,15 +1,23 @@
 #include "Thetis_BNO055.h"
 
-bool initBNO055(Adafruit_BNO055 &imu, Stream &out) {
-    out.print("Initializing BNO055...");
+Adafruit_BNO055 bno;
+
+bool initBNO055() {
+    #ifdef BNO055_DEBUG
+    DEBUG_SERIAL_PORT.print("Initializing BNO055...");
+    #endif
     digitalWrite(BNO055_RESET, HIGH); // Ensure BNO Reset is HIGH
-    if (!imu.begin()) {
-        out.println("Failed to initialize BNO055");
+    if (!bno.begin()) {
+        #ifdef BNO055_DEBUG
+        DEBUG_SERIAL_PORT.println("Failed to initialize BNO055");
+        #endif
         return false;
     }
     else {
-        imu.setExtCrystalUse(true);
-        out.println("done!");
+        bno.setExtCrystalUse(true);
+        #ifdef BNO055_DEBUG
+        DEBUG_SERIAL_PORT.println("done!");
+        #endif
         return true;
     }
 }